Refactored, added packages
