+++ homework-1/converter/converter.go
@@ -1,12 +1,13 @@
-package main
+package converter
 
 import "fmt"
 
-func main() {
+func Calculate() {
 	const dollarRate float64 = 76.26
 	var rub float64
 	fmt.Println("Введите количество рублей для обмена:")
 	fmt.Scanln(&rub)
 	sum := rub / dollarRate
-	fmt.Printf("%.2f", sum)
+	fmt.Print("Вы получите: ")
+	fmt.Printf("$%.2f\n\n", sum)
 }
